Assignment05/chatbot-litellm/pyproject.toml
[project]
name = "chatbot-litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.71.1",
    "openai-agents[litellm]>=0.0.16",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
